pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cdc-platform"
version = "0.1.0"
description = "Change Data Capture platform"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "confluent-kafka[avro]>=2.6,<3",
    "httpx>=0.28,<1",
    "tenacity>=9,<10",
    "structlog>=24,<26",
    "typer>=0.15,<1",
    "pyyaml>=6,<7",
    "rich>=13,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6,<7",
    "respx>=0.22,<1",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "psycopg2-binary>=2.9,<3",
]

[project.scripts]
cdc = "cdc_platform.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cdc_platform"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring Docker services",
]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.pycodestyle]
max-line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
